Resourceコマンドにフィルタを追加
