+++ src/shared/utility/system_info.ts
@@ -22,7 +22,7 @@ export const SystemInfo = {
     name: "AntOS",
   },
   application: {
-    version: "9.2.96",
+    version: "9.2.98",
     name: "Mitsuyoshi bot",
   },
 }
